feat: refactor nemo_tract into sub-modules
