--- torch_to_nnef/nemo_tract.py
@@ -1,1759 +1,0 @@
-"""Support for NVIDIA NeMo models export to NNEF (with TractNNEF focus).
-
-Provide utilities to export NeMo models, particularly ASR models,
-to the NNEF format using TractNNEF.
-Includes functions to handle model subnets, dynamic axes, and
-custom extensions required for the export process.
-
-"""
-
-import argparse
-from collections import OrderedDict
-import datetime
-import json
-import logging
-import sys
-import typing as T
-from contextlib import contextmanager
-from dataclasses import dataclass
-from pathlib import Path
-
-import torch
-
-from torch_to_nnef._optional_types import (
-    HuggingFaceHubModule,
-    InjectedHuggingFaceHubModule,
-    InjectedLightningModule,
-    InjectedNemoModule,
-    InjectedOmegaConfModule,
-    InjectedQuestionaryModule,
-)
-from torch_to_nnef.compress import (
-    DEFAULT_COMPRESSION_REGISTRY,
-    dynamic_load_registry,
-)
-from torch_to_nnef.export import export_model_to_nnef
-from torch_to_nnef.inference_target.base import InferenceTarget
-from torch_to_nnef.inference_target.tract import (
-    TractCheckTolerance,
-    TractCli,
-    TractNNEF,
-    build_io,
-)
-from torch_to_nnef.log import init_log, set_lib_log_level
-from torch_to_nnef.torch_graph.ir_naming import VariableNamingScheme
-from torch_to_nnef.utils import (
-    INJECTED,
-    SemanticVersion,
-    T2NExtra,
-    require_extra_decorator,
-)
-
-# https://huggingface.co/nvidia/parakeet-tdt-0.6b-v3
-PARAKEET_V3_SLUG = "nvidia/parakeet-tdt-0.6b-v3"
-PARAKEET_110M_SLUG = "parakeet-tdt_ctc-110m"
-# https://huggingface.co/nvidia/nemotron-speech-streaming-en-0.6b
-NEMOTRON_0_6B = "nvidia/nemotron-speech-streaming-en-0.6b"
-
-
-LOGGER = logging.getLogger(__name__)
-
-
-@contextmanager
-@require_extra_decorator(extra=T2NExtra.NEMO_TRACT, module="nemo")
-@require_extra_decorator(extra=T2NExtra.NEMO_TRACT, module="pytorch_lightning")
-def exportable_nemo_net(
-    output_name,
-    model,
-    input_example,
-    use_dynamo=False,
-    batch_size: int = 1,
-    float_dtype: T.Optional[torch.dtype] = None,
-    *,
-    nemo: InjectedNemoModule = INJECTED,
-    pytorch_lightning: InjectedLightningModule = INJECTED,
-):
-    """Context manager to follow export way of nemo models.
-
-    see: nemo.core.classes.Exportable._export
-    """
-    typecheck = nemo.core.classes.typecheck
-    exportable_class = nemo.core.classes.exportable.Exportable
-    parse_input_example = nemo.utils.export_utils.parse_input_example
-    wrap_forward_method = nemo.utils.export_utils.wrap_forward_method
-    my_args = {"use_dynamo": use_dynamo}
-
-    model.eval()
-    for param in model.parameters():
-        param.requires_grad = False
-
-    exportables = []
-    for m in model.modules():
-        if isinstance(m, exportable_class):
-            exportables.append(m)
-
-    forward_method = None
-    old_forward_method = None
-    try:
-        # Disable typechecks
-        typecheck.set_typecheck_enabled(enabled=False)
-
-        # Allow user to completely override forward method to export
-        forward_method, old_forward_method = wrap_forward_method(model)
-
-        # Set module mode
-        with (
-            torch.inference_mode(),
-            torch.no_grad(),
-            torch.jit.optimized_execution(True),
-            pytorch_lightning.core.module._jit_is_scripting(),
-        ):
-            imod = model
-            if hasattr(imod, "input_module"):
-                imod = model.input_module
-            if input_example is None:
-                if float_dtype is None:
-                    try:
-                        fdtype = next(imod.parameters()).dtype
-                    except StopIteration:
-                        fdtype = torch.float32
-                else:
-                    fdtype = float_dtype
-                LOGGER.debug("Generating dummy input... %s", float_dtype)
-                input_example = imod.input_example(max_batch=batch_size)
-                # Cast to correct dtype (usualy float16 if not float16)
-                if fdtype != torch.float32:
-                    input_example = [
-                        ie.to(fdtype)
-                        if isinstance(ie, torch.Tensor)
-                        and ie.dtype == torch.float32
-                        else ie
-                        for ie in input_example
-                    ]
-
-            # Run (posibly overridden) prepare methods before calling forward()
-            for ex in exportables:
-                if hasattr(ex, "_prepare_for_export"):
-                    ex._prepare_for_export(**my_args, noreplace=True)
-
-            if hasattr(model, "_prepare_for_export"):
-                model._prepare_for_export(
-                    output=output_name, input_example=input_example, **my_args
-                )
-
-            input_list, input_dict = parse_input_example(input_example)
-            output_example = model.forward(*input_list, **input_dict)
-            if not isinstance(output_example, tuple):
-                output_example = (output_example,)
-
-            # dynamic axis is a mapping from input/output_name
-            # => list of "dynamic" indices
-            dynamic_axes = model.dynamic_shapes_for_export(use_dynamo)
-            yield input_example, output_example, dynamic_axes
-    finally:
-        typecheck.enable_wrapping(enabled=True)
-        typecheck.set_typecheck_enabled(enabled=True)
-        if forward_method:
-            type(model).forward = old_forward_method
-        if hasattr(model, "_export_teardown"):
-            model._export_teardown()
-
-
-def decoder_fix_input_example_batch_size(
-    input_example: T.Tuple[torch.Tensor, ...],
-    batch_size: int,
-) -> T.List[torch.Tensor]:
-    """Fix the batch size of the input example for decoder models.
-
-    NeMo decoder batch size option input example is wrong.
-    This function adjusts the input example to have the specified batch size.
-
-    Args:
-        input_example: The original input example tuple (input_ids, ...).
-        batch_size: The desired batch size.
-
-    Returns:
-        The adjusted input example tuple with the specified batch size.
-    """
-
-    def expand_dim(x, dim, size):
-        shape = list(x.shape)
-        shape[dim] = size
-        return x.expand(*shape)
-
-    assert batch_size > 0, "Batch size must be positive."
-    input_ids, *encoder_outputs = input_example
-    if batch_size != 1 and input_ids.size(0) == 1:
-        input_ids = expand_dim(input_ids, 0, batch_size)
-    return [input_ids] + list(encoder_outputs)
-
-
-@require_extra_decorator(extra=T2NExtra.NEMO_TRACT, module="nemo")
-def iter_nemo_model_subnets(
-    model,
-    input_example=None,
-    float_dtype: T.Optional[torch.dtype] = None,
-    split_joint_decoder: bool = False,
-    remove_unused_inputs: bool = True,
-    apply_sequential_examples: bool = False,
-    batch_size: int = 3,
-    *,
-    nemo: InjectedNemoModule = INJECTED,
-):
-    """Iterator over exportable subnets of a nemo model.
-
-    Args:
-        model: NeMo model to iterate over.
-        input_example: Optional input example to use for export.
-        float_dtype: Optional float dtype to use for export.
-        split_joint_decoder: To split joint decoder subnets (if encountered).
-        remove_unused_inputs: To remove unused inputs from subnet exports.
-        apply_sequential_examples: If True, use sequential input examples
-            for each subnet.
-        batch_size: Batch size to use for dummy input generation.
-
-    Yields:
-        subnet_name: name of the subnet
-        subnet: the subnet module
-        input_example: input example for the subnet
-        dynamic_axes: dynamic axes info for the subnet
-
-
-    see: nemo.core.classes.Exportable.export
-
-    """
-    subnet_names = model.list_export_subnets()
-    allow_same_io_names = [False] * len(subnet_names)
-    nemo_model_mod = nemo.collections.asr.models
-    if isinstance(
-        model, nemo_model_mod.classification_models.EncDecClassificationModel
-    ):
-        subnet_names = ["encoder", "decoder"]
-        allow_same_io_names = [True, False]
-
-        # Get the class you want to patch
-        cls = model.encoder.__class__
-
-        # Capture original property getter BEFORE patching
-        orig_fget = cls.output_types.fget  # <-- important
-
-        def patched_output_types(self):
-            original = orig_fget(self)  # call original getter
-
-            new = OrderedDict()
-            for k, v in original.items():
-                if k == "encoded_lengths":
-                    new["length"] = v
-                else:
-                    new[k] = v
-            return new
-
-        # Patch the right attribute name
-        cls.output_types = property(patched_output_types)
-
-    # Handle CTC families (EncDecCTCModel and EncDecCTCModelBPE)
-    EncDecCTCModel = None
-    EncDecCTCModelBPE = None
-    # Newer NeMo often exposes ctc_models submodule
-    try:
-        EncDecCTCModel = getattr(
-            nemo_model_mod.ctc_models, "EncDecCTCModel", None
-        )
-        EncDecCTCModelBPE = getattr(
-            nemo_model_mod.ctc_models, "EncDecCTCModelBPE", None
-        )
-    except AttributeError:  # pragma: no cover - defensive
-        pass
-    # Fallback: sometimes classes are directly under models
-    if EncDecCTCModel is None:
-        EncDecCTCModel = getattr(nemo_model_mod, "EncDecCTCModel", None)
-    if EncDecCTCModelBPE is None:
-        EncDecCTCModelBPE = getattr(nemo_model_mod, "EncDecCTCModelBPE", None)
-
-    if (
-        (EncDecCTCModel is not None and isinstance(model, EncDecCTCModel))
-        or (
-            EncDecCTCModelBPE is not None
-            and isinstance(model, EncDecCTCModelBPE)
-        )
-    ):
-        subnet_names = ["encoder", "decoder"]
-        allow_same_io_names = [True, False]
-        auto_seq = True
-
-        # Patch encoder output_types to map encoded_lengths -> length
-        cls = model.encoder.__class__
-        try:
-            orig_fget = cls.output_types.fget  # type: ignore[attr-defined]
-        except AttributeError:  # pragma: no cover - defensive
-            orig_fget = None
-
-        def patched_output_types_ctc(self):
-            original = (
-                orig_fget(self)
-                if orig_fget is not None
-                else getattr(self, "output_types", {})
-            )
-            try:
-                items = original.items()
-            except Exception:  # pragma: no cover - defensive
-                return original
-            new = OrderedDict()
-            for k, v in items:
-                new["length" if k == "encoded_lengths" else k] = v
-            return new
-
-        try:
-            cls.output_types = property(patched_output_types_ctc)  # type: ignore[attr-defined]
-        except Exception:  # pragma: no cover - defensive
-            pass
-
-    for subnet_name, sio in zip(subnet_names, allow_same_io_names):
-        subnet = model.get_export_subnet(subnet_name)
-        if subnet_name == "decoder_joint":
-            input_example = None  # reset input example for joint
-            # because need more parameters than encoder output only
-        with exportable_nemo_net(
-            subnet_name,
-            subnet,
-            input_example,
-            batch_size=batch_size,
-            float_dtype=float_dtype,
-        ) as (
-            #  pylint: disable-next=redefined-argument-from-local
-            input_example,
-            out_example,
-            dynamic_axes,
-        ):
-            if subnet_name == "decoder_joint":
-                if split_joint_decoder:
-                    # split into decoder and joint
-                    # inputs are force generated by nemo .input_example()
-                    decoder = subnet.decoder
-                    if remove_unused_inputs:
-                        decoder = DecoderWithoutTargetLength(decoder)
-                    yield (
-                        "decoder",
-                        decoder,
-                        decoder_fix_input_example_batch_size(
-                            decoder.input_example(max_batch=batch_size),
-                            batch_size=batch_size,
-                        ),
-                        decoder.dynamic_shapes_for_export(False),
-                        sio,
-                    )
-                    yield (
-                        "joint",
-                        subnet.joint,
-                        subnet.joint.input_example(max_batch=batch_size),
-                        subnet.joint.dynamic_shapes_for_export(False),
-                        sio,
-                    )
-                    continue
-                if remove_unused_inputs:
-                    subnet = DecoderWithoutTargetLength(subnet)
-                    input_example = subnet.filter_original_input_example(
-                        input_example
-                    )
-                input_example = decoder_fix_input_example_batch_size(
-                    input_example,
-                    batch_size=batch_size,
-                )
-            if len(input_example) > len(subnet.input_names):
-                # if < that means some inputs are optional
-                raise RuntimeError(
-                    "declared input names:",
-                    subnet.input_names,
-                    f"but expected {len(input_example)} inputs",
-                )
-            yield subnet_name, subnet, input_example, dynamic_axes, sio
-            # Propagate input example
-            # (default scenario, may need to be overriden)
-            if input_example is not None and apply_sequential_examples:
-                input_example = out_example
-            else:
-                input_example = None
-
-
-def build_dynamic_axes(subnet, nemo_dynamic_axes):  # noqa: MC0001
-    """Build dynamic axes mapping and custom extensions for nemo subnet.
-
-    Args:
-        subnet: nemo subnet module
-        nemo_dynamic_axes: dynamic axes info from nemo export
-    Returns:
-        dynamic_axes: dynamic axes mapping for torch_to_nnef
-        custom_extensions: custom extensions for torch_to_nnef
-
-    Note:
-        this code will not scale well and should be refactored when more
-        nemo models are supported.
-    """
-    dynamic_axes = {}
-    # Assume each input always start by Batch dimension
-    custom_extensions = set()
-
-    def build_partial_dynamic_axes(
-        iname: str, symbols: T.Union[str, T.List[str]], suffix: str = ""
-    ):
-        siname = iname + suffix
-        dynamic_axes[siname] = {}
-        for axis in nemo_dynamic_axes[iname]:
-            if symbols[axis] in "BSA":
-                custom_extensions.add(f"tract_assert {symbols[axis]} >= 1")
-            dynamic_axes[siname][axis] = symbols[axis]
-
-    for iname in subnet.input_names:
-        if iname in nemo_dynamic_axes:
-            if not nemo_dynamic_axes[iname]:
-                continue
-            symbols = ""
-            if iname == "input_signal":
-                symbols = "BA"  # Batch, audio Frames
-            elif iname == "audio_signal":
-                assert max(nemo_dynamic_axes[iname]) < 3
-                symbols = "BFS"  # Batch, Features, Stream
-            elif iname == "length":
-                assert max(nemo_dynamic_axes[iname]) < 1
-                symbols = "B"  # Batch
-            elif iname == "encoder_outputs":
-                # example: B,512,(S+7)/8,F32
-                symbols = "BHR"  # Batch, High End Features, ReducedStream
-            elif iname == "targets":
-                symbols = "BT"  # Batch, TargetInfo
-            elif iname == "target_length":
-                symbols = "B"  # Batch
-            elif iname == "input_states_1":
-                symbols = ["STATES_1_DIM_1", "B", "STATES_1_DIM_2"]
-            elif iname == "input_states_2":
-                symbols = ["STATES_2_DIM_1", "B", "STATES_2_DIM_2"]
-            elif iname == "states":
-                build_partial_dynamic_axes(
-                    iname,
-                    ["STATES_1_DIM_1", "B", "STATES_1_DIM_2"],
-                    suffix="_0",
-                )
-                build_partial_dynamic_axes(
-                    iname,
-                    ["STATES_2_DIM_1", "B", "STATES_2_DIM_2"],
-                    suffix="_1",
-                )
-                continue
-            elif iname == "decoder_outputs":
-                # Batch, output decoder, Unknown, Time dimension decoder
-                symbols = "BOUT"
-            elif iname == "encoder_output":
-                # example: B,512,(S+7)/8,F32
-                symbols = "BHR"  # Batch, High End Features, ReducedStream
-            else:
-                raise NotImplementedError(
-                    f"cannot guess dynamic axis symbols for input '{iname}'"
-                )
-            build_partial_dynamic_axes(iname, symbols)
-    return dynamic_axes, custom_extensions
-
-
-def _collapse_axes_for_input(
-    orig_axes: T.Dict[int, str],
-    *,
-    full_axes_spec: T.Optional[T.Sequence[str]] = None,
-    assume_batch_at0: bool = True,
-) -> T.Dict[int, str]:
-    """Collapse a single input axes mapping by removing 'B' and reindexing.
-
-    Supports cases where batch is not reported as dynamic by inferring it from
-    the full static axes spec or assuming B at index 0.
-    """
-    # Determine batch positions
-    if full_axes_spec is not None:
-        b_positions = [ix for ix, sym in enumerate(full_axes_spec) if sym == "B"]
-    else:
-        pairs_for_b = sorted(orig_axes.items(), key=lambda kv: kv[0])
-        b_positions = [ix for ix, sym in pairs_for_b if sym == "B"]
-        if assume_batch_at0 and 0 not in b_positions:
-            b_positions = [0] + b_positions
-
-    pairs = sorted(orig_axes.items(), key=lambda kv: kv[0])
-    new_axes: T.Dict[int, str] = {}
-    for orig_ix, sym in pairs:
-        if sym == "B":
-            continue
-        n_b_before = sum(1 for b in b_positions if b < orig_ix)
-        new_ix = orig_ix - n_b_before
-        new_axes[new_ix] = sym
-    return new_axes
-
-
-def collapse_dynamic_axes_mapping(
-    nemo_dynamic_axes: T.Dict[str, T.Dict[int, str]],
-    keep_input_names: T.Sequence[str],
-    *,
-    full_axes_by_name: T.Optional[T.Dict[str, T.Sequence[str]]] = None,
-) -> T.Dict[str, T.Dict[int, str]]:
-    """Produce a collapsed dynamic axis mapping without batch dims.
-
-    - Keeps only inputs present in keep_input_names.
-    - Removes 'B' from the axis symbols and compacts indices.
-    """
-    collapsed: T.Dict[str, T.Dict[int, str]] = {}
-    for name in keep_input_names:
-        axes = nemo_dynamic_axes.get(name)
-        if not axes:
-            continue
-        collapsed[name] = _collapse_axes_for_input(
-            axes,
-            full_axes_spec=(full_axes_by_name or {}).get(name),
-        )
-    return collapsed
-
-
-ExportParameters = T.NamedTuple(
-    "ExportParameters",
-    [
-        ("name", str),
-        ("model", torch.nn.Module),
-        ("test_input", object),
-        ("inference_target", InferenceTarget),
-        ("input_names", list),
-        ("output_names", list),
-        ("custom_extensions", list),
-        ("allow_same_io_names", bool),
-        ("specific_tract_properties", dict),
-    ],
-)
-
-
-@require_extra_decorator(extra=T2NExtra.NEMO_TRACT, module="nemo")
-def build_custom_subnet_tract_properties(
-    subnet_name, subnet, *, nemo: InjectedNemoModule = INJECTED
-):
-    """Build custom tract properties for nemo subnet."""
-    return {
-        "subnet_name": subnet_name,
-        "n_parameters": sum(_.numel() for _ in subnet.parameters()),
-        "nemo_version": nemo.__version__,
-    }
-
-
-@require_extra_decorator(extra=T2NExtra.NEMO_TRACT, module="nemo")
-def iter_export_params_for_generic_nemo_asr_model(
-    asr_model,
-    inference_target,
-    skip_preprocessor: bool = False,
-    split_joint_decoder: bool = False,
-    remove_unused_inputs: bool = True,
-    collapse_batch_dim: bool = False,
-    float_dtype: T.Optional[torch.dtype] = None,
-    *,
-    nemo: InjectedNemoModule = INJECTED,
-) -> T.Iterator[ExportParameters]:
-    """Iterator over export parameters for a generic NeMo ASR model.
-
-    Args:
-        asr_model: The NeMo ASR model to export.
-        inference_target: The target inference type.
-        skip_preprocessor: Whether to skip exporting the preprocessor subnet.
-        split_joint_decoder:
-            Whether to split the joint and decoder subnets exported.
-        remove_unused_inputs:
-            Whether to remove unused inputs from the exported model.
-        float_dtype: Optional float dtype to use for export.
-
-    Yields:
-        ExportParameters for each subnet of the ASR model, with the preprocessor
-    """
-    asr_model.eval()
-
-    if not skip_preprocessor:
-        inps = asr_model.preprocessor.input_example()
-        if hasattr(asr_model.preprocessor, "featurizer") and hasattr(
-            asr_model.preprocessor.featurizer, "dither"
-        ):
-            # disable dither for export
-            if asr_model.preprocessor.featurizer.dither != 0.0:
-                LOGGER.info("disabling dither for preprocessor export")
-            asr_model.preprocessor.featurizer.dither = 0.0
-        if hasattr(asr_model.preprocessor, "featurizer") and hasattr(
-            asr_model.preprocessor.featurizer, "pad_to"
-        ):
-            if asr_model.preprocessor.featurizer.pad_to != 0.0:
-                LOGGER.info("disabling pad_to for preprocessor export")
-            asr_model.preprocessor.featurizer.pad_to = 0
-
-        if isinstance(
-            asr_model.preprocessor,
-            nemo.collections.asr.modules.audio_preprocessing.AudioPreprocessor,
-        ):
-            asr_model.preprocessor = WrapAudioPreprocessor(
-                asr_model.preprocessor
-            )
-
-        with exportable_nemo_net(
-            "preprocessor", asr_model.preprocessor, inps
-        ) as (
-            input_example,
-            _,
-            nemo_dynamic_axes,
-        ):
-            dynamic_axes, custom_extensions = build_dynamic_axes(
-                asr_model.preprocessor, nemo_dynamic_axes
-            )
-
-            subnet_name = "preprocessor"
-            model = asr_model.preprocessor
-            input_names = model.input_names[: len(input_example)]
-            output_names = model.output_names
-            test_input = inps or input_example
-            dyn = dynamic_axes
-            if collapse_batch_dim:
-                # Wrap and collapse axes
-                model = CollapseBatchDimWrapper(model, dynamic_axes)
-                input_names = model.input_names
-                output_names = model.output_names
-                test_input = model.input_example()
-                dyn = collapse_dynamic_axes_mapping(dynamic_axes, input_names)
-
-            yield ExportParameters(
-                name=subnet_name,
-                model=model,
-                test_input=test_input,
-                inference_target=inference_target.with_dynamic_axes(dyn),
-                input_names=input_names,
-                output_names=output_names,
-                custom_extensions=list(custom_extensions),
-                allow_same_io_names=False,  # not used for preprocessor export
-                specific_tract_properties=build_custom_subnet_tract_properties(
-                    subnet_name, model
-                ),
-            )
-
-    for (
-        subnet_name,
-        subnet,
-        input_example,
-        nemo_dynamic_axes,
-        allow_same_io_names,
-    ) in iter_nemo_model_subnets(
-        asr_model,
-        float_dtype=float_dtype,
-        split_joint_decoder=split_joint_decoder,
-        remove_unused_inputs=remove_unused_inputs,
-    ):
-        dynamic_axes, custom_extensions = build_dynamic_axes(
-            subnet, nemo_dynamic_axes
-        )
-
-        model = subnet
-        test_input = input_example
-        input_names = subnet.input_names[: len(input_example)]
-        output_names = subnet.output_names
-        # Limit dynamic axes to the inputs we are actually exposing
-        dyn = {k: v for k, v in dynamic_axes.items() if k in input_names}
-
-        if collapse_batch_dim:
-            model = CollapseBatchDimWrapper(subnet, dynamic_axes)
-            test_input = model.input_example()
-            input_names = model.input_names
-            output_names = model.output_names
-            # Use wrapper's collapsed dynamic mapping for correctness
-            dyn = model.dynamic_shapes_for_export()
-
-        yield ExportParameters(
-            name=subnet_name,
-            model=model,
-            test_input=test_input,
-            inference_target=inference_target.with_dynamic_axes(dyn),
-            input_names=input_names,
-            output_names=output_names,
-            custom_extensions=list(custom_extensions),
-            allow_same_io_names=allow_same_io_names,
-            specific_tract_properties=build_custom_subnet_tract_properties(
-                subnet_name, model
-            ),
-        )
-
-
-@require_extra_decorator(extra=T2NExtra.NEMO_TRACT, module="omegaconf")
-def export_nemo_asr_model(
-    asr_model,
-    inference_target,
-    export_dir: Path,
-    compress_registry: str,
-    compress_method: T.Optional[str] = None,
-    skip_preprocessor: bool = False,
-    split_joint_decoder: bool = False,
-    extra_cfg: T.Optional[T.Dict[str, T.Any]] = None,
-    float_dtype: T.Optional[torch.dtype] = None,
-    remove_unused_inputs: bool = True,
-    dump_checked_io: bool = False,
-    collapse_batch_dim: bool = False,
-    *,
-    omegaconf: InjectedOmegaConfModule = INJECTED,
-    **kwargs,
-):
-    """Export a generic NeMo ASR model to NNEF format using TractNNEF.
-
-    Args:
-        asr_model: The NeMo ASR model to export.
-        inference_target: The inference target configuration for export.
-        export_dir: Directory where the exported NNEF files will be saved.
-        skip_preprocessor: If True, skip exporting the preprocessor subnet.
-        split_joint_decoder: Whether to split the joint&decoder subnets export.
-        compress_registry: Compression registry for the exported NNEF subnets.
-        compress_method: Compression method for the exported NNEF subnets.
-            if None, no compression is applied.
-        extra_cfg: Additional configuration to save alongside the model.
-        float_dtype: Optional float dtype to use for export.
-        remove_unused_inputs: To remove unused inputs in the exported model.
-            This happen for decoder subnets that do not use target_length.
-        dump_checked_io: Whether to dump checked input/output examples.
-        omegaconf: Injected OmegaConf module.
-        kwargs: Additional keyword arguments to pass to the export function.
-    """
-    with (export_dir / "model_config.json").open("w", encoding="utf8") as fh:
-        cfg = omegaconf.OmegaConf.to_container(asr_model.cfg)
-        if extra_cfg is not None:
-            cfg.update(extra_cfg)
-        json.dump(cfg, fh, indent=2)
-    if compress_method:
-        LOGGER.info("use compresssion: %s", compress_method)
-        registry = dynamic_load_registry(compress_registry)
-        asr_model = registry[compress_method](
-            asr_model,
-            export_dirpath=export_dir,
-        )
-        LOGGER.info("successfully applied compression: %s", compress_method)
-
-    for export_params in iter_export_params_for_generic_nemo_asr_model(
-        asr_model,
-        inference_target,
-        skip_preprocessor=skip_preprocessor,
-        split_joint_decoder=split_joint_decoder,
-        float_dtype=float_dtype,
-        remove_unused_inputs=remove_unused_inputs,
-        collapse_batch_dim=collapse_batch_dim,
-    ):
-        LOGGER.info("start subnet export: %s", export_params.name)
-        if dump_checked_io:
-            test_dir = export_dir / "test"
-            test_dir.mkdir(parents=True, exist_ok=True)
-            build_io(
-                export_params.model,
-                export_params.test_input,
-                io_npz_path=test_dir / f"{export_params.name}_checked_io.npz",
-                input_names=export_params.input_names,
-                output_names=export_params.output_names,
-            )
-        export_model_to_nnef(
-            model=export_params.model,
-            args=export_params.test_input,
-            inference_target=export_params.inference_target.with_specific_properties(
-                export_params.specific_tract_properties
-            ),
-            input_names=export_params.input_names,
-            output_names=export_params.output_names,
-            file_path_export=export_dir / f"{export_params.name}.nnef.tgz",
-            custom_extensions=export_params.custom_extensions,
-            allow_same_io_names=export_params.allow_same_io_names,
-            **kwargs,
-        )
-        LOGGER.info("exported subnet: %s with success", export_params.name)
-
-
-def parser_cli():
-    """Build the CLI parser for NeMo ASR model export to NNEF format."""
-    parser = argparse.ArgumentParser(
-        description="Export NeMo ASR model to NNEF format using TractNNEF."
-    )
-    parser.add_argument(
-        "-s",
-        "--model-slug",
-        type=str,
-        default="*",
-        help="The model slug for the NeMo ASR model to export."
-        "if you don't know just live it blank (select box will be proposed)",
-    )
-    parser.add_argument(
-        "-e",
-        "--export-dir",
-        type=Path,
-        required=True,
-        help="Directory to save the exported NNEF files.",
-    )
-    parser.add_argument(
-        "--skip-preprocessor",
-        action="store_true",
-        help="Skip exporting the preprocessor subnet.",
-    )
-    parser.add_argument(
-        "--split-joint-decoder",
-        action="store_true",
-        help="Split the joint and decoder subnets during export.",
-    )
-    parser.add_argument(
-        "--force-sdpa-pytorch",
-        action="store_true",
-        help="Forcing sdpa to use PyTorch implementation."
-        " (likely more efficent, once stable tract side)",
-    )
-    parser.add_argument(
-        "--collapse-batch-dim",
-        action="store_true",
-        help=(
-            "Remove batch dimension from exported subnet interfaces and "
-            "hide batch-only length inputs (length, target_length, ...)."
-        ),
-    )
-    parser.add_argument(
-        "--tract-reify-sdpa",
-        action="store_true",
-        help="Forcing sdpa to be reified in NNEF. "
-        " Works only if --force-sdpa-pytorch is set."
-        " this is auto activated if tract version used is "
-        " at least 0.23.0 which support stable sdpa reifycation natively.",
-    )
-    parser.add_argument(
-        "-dt",
-        "--data-type",
-        type=str,
-        choices=["float32", "float16", "mixed"],
-        help="Data of most weights for export (experimental).",
-    )
-    parser.add_argument(
-        "-n",
-        "--naming-scheme",
-        default=VariableNamingScheme.NATURAL_VERBOSE_CAMEL.value,
-        choices=[vns.value for vns in VariableNamingScheme],
-        help="display debug information",
-    )
-    parser.add_argument(
-        "--tract-specific-path",
-        required=False,
-        help="tract specific path (instead of latest version)",
-    )
-    parser.add_argument(
-        "--tract-specific-version",
-        required=False,
-        help="tract specific version",
-    )
-    parser.add_argument(
-        "-tt",
-        "--tract-check-io-tolerance",
-        default=TractCheckTolerance.APPROXIMATE.value,
-        choices=[t.value for t in TractCheckTolerance] + ["skip"],
-        help="tract check io tolerance level",
-    )
-
-    parser.add_argument(
-        "--compress-registry",
-        type=str,
-        default=DEFAULT_COMPRESSION_REGISTRY,
-        help="compression registry for the exported nnef subnets",
-    )
-
-    parser.add_argument(
-        "--compress-method",
-        type=str,
-        default=None,
-        help="compression method for the exported nnef subnets",
-    )
-
-    parser.add_argument(
-        "--dump-checked-io",
-        required=False,
-        default=False,
-        action="store_true",
-        help="dump tested io to the given path for checking purpose",
-    )
-
-    parser.add_argument(
-        "-v",
-        "--verbose",
-        action="store_true",
-        help="display debug information",
-    )
-
-    return parser.parse_args()
-
-
-def setup_inference_target_from_cli_args(args) -> TractNNEF:
-    """Setup TractNNEF inference target from CLI arguments."""
-    if args.tract_specific_version:
-        assert args.tract_specific_path is None, "set either version or path"
-        inference_target = TractNNEF(
-            SemanticVersion.from_str(args.tract_specific_version)
-            if isinstance(args.tract_specific_version, str)
-            else args.tract_specific_version
-        )
-    elif args.tract_specific_path:
-        tract_cli_path = Path(args.tract_specific_path)
-        assert tract_cli_path.exists(), tract_cli_path
-        tract_cli = TractCli(tract_cli_path)
-        inference_target = TractNNEF(
-            tract_cli.version,
-            specific_tract_binary_path=tract_cli_path,
-        )
-    else:
-        inference_target = TractNNEF.latest()
-    if args.tract_check_io_tolerance == "skip":
-        inference_target.check_io = False
-    else:
-        inference_target.check_io_tolerance = args.tract_check_io_tolerance
-
-    if args.tract_reify_sdpa:
-        inference_target.reify_sdpa_operator = True
-        if not args.force_sdpa_pytorch and inference_target.version < "0.23.0":
-            LOGGER.warning(
-                "Reifying sdpa without forcing pytorch implementation "
-                "will likely export no sdpa operators since not expressed "
-                "in models. It is recommended to use --force-sdpa-pytorch "
-                "when using --tract-reify-sdpa, prior to tract 0.23.0 release "
-                "where sdpa reifycation is activated by default."
-            )
-    return inference_target
-
-
-class WrapAudioPreprocessor(torch.nn.Module):
-    """Wraps the AudioPreprocessor to fix input_example empty."""
-
-    def __init__(self, preprocessor: torch.nn.Module):
-        super().__init__()
-        self.preprocessor = preprocessor
-
-    def input_example(self, max_batch: int = 1):
-        # return a dummy input example if the original is empty
-        ie = self.preprocessor.input_example()
-        if ie is None or len(ie) == 0:
-            # use self.preprocessor.input_types
-            # {'input_signal': NeuralType(axis=(batch, time), element_type=AudioSignal), 'length': NeuralType(axis=(batch,), element_type=LengthsType)}
-            # to build tensor dummy input example, with 16000 frames (1 second of audio at 16kHz)
-
-            input_types = self.preprocessor.input_types
-            batch_size = max_batch
-            default_time = 16000  # safe default for time axis
-
-            example = []
-
-            for _, neural_type in input_types.items():
-                axes = neural_type.axes  # tuple of axis descriptors
-
-                shape = []
-                dtype = torch.float32  # default
-
-                for axis in axes:
-                    # Axis may be a string or AxisType depending on NeMo version
-                    axis_name = getattr(axis, "kind", axis)
-                    axis_name = str(axis_name).lower()
-
-                    if "batch" in axis_name or axis_name == "b":
-                        shape.append(batch_size)
-                    elif "time" in axis_name or axis_name == "t":
-                        shape.append(default_time)
-                    else:
-                        # Fallback: set to 1 for unknown dims
-                        shape.append(1)
-
-                # LengthsType → long tensor
-                element_type_name = type(
-                    neural_type.elements_type
-                ).__name__.lower()
-                if "length" in element_type_name:
-                    dtype = torch.long
-                    tensor = torch.full(
-                        (batch_size,),
-                        default_time,
-                        dtype=dtype,
-                    )
-                else:
-                    tensor = torch.zeros(
-                        *shape,
-                        dtype=dtype,
-                    )
-
-                example.append(tensor)
-
-            return tuple(example)
-        return ie
-
-    def dynamic_shapes_for_export(self, *args, **kwargs):
-        name_map = {
-            "batch": "B",
-            "time": "T",
-            "stream": "S",
-        }
-        return {
-            k: {ix: name_map[str(ax.kind)] for ix, ax in enumerate(v.axes)}
-            for k, v in self.preprocessor.input_types.items()
-        }
-
-    @property
-    def input_names(self):
-        return list(self.preprocessor.input_types.keys())
-
-    @property
-    def output_names(self):
-        return list(self.preprocessor.output_types.keys())
-
-    def forward(self, *args, **kwargs):
-        return self.preprocessor(*args, **kwargs)
-
-
-class WrapPreprocessorCast(torch.nn.Module):
-    """Wraps the preprocessor to add a cast to float32 at the output."""
-
-    def __init__(self, preprocessor: torch.nn.Module, dtype: torch.dtype):
-        super().__init__()
-        self.preprocessor = preprocessor
-        self.dtype = dtype
-
-    def forward(self, *args, **kwargs):
-        x = self.preprocessor(*args, **kwargs)
-        return tuple([x[0].to(self.dtype)] + list(x)[1:])
-
-    def input_example(self):
-        return self.preprocessor.input_example()
-
-    def _export_teardown(self):
-        self.preprocessor._export_teardown()
-
-    def _prepare_for_export(self, *args, **kwargs):
-        self.preprocessor._prepare_for_export(*args, **kwargs)
-
-    def dynamic_shapes_for_export(self, *args, **kwargs):
-        return self.preprocessor.dynamic_shapes_for_export(*args, **kwargs)
-
-    @property
-    def input_names(self):
-        return self.preprocessor.input_names
-
-    @property
-    def output_names(self):
-        return self.preprocessor.output_names
-
-
-class DecoderWithoutTargetLength(torch.nn.Module):
-    """Wraps the decoder or joint+decoder for export.
-
-    This remove the parameters 'target_length' that are
-    not needed during inference...
-
-    Enabled classes:
-        `nemo.collections.asr.modules.rnnt.RNNTDecoderJoint`
-        `nemo.collections.asr.modules.rnnt.RNNTDecoder`
-
-    Alter forward by auto adding the target_length parameter
-    based on the shape of the input tensors (Batch size).
-    as an array of shape (batch_size, 1) full of ones.
-    Then remove it from the output (this is the 2nd argument).
-    This is only applied for enabled classes.
-
-    This should lead at export to a complete removal
-    of the unused target_length.
-
-    """
-
-    FILTER_ARGUMENT = "target_length"
-    FILTER_OUTPUT = "prednet_lengths"
-
-    @require_extra_decorator(
-        extra=T2NExtra.NEMO_TRACT, module="nemo.collections.asr", kw="nemo_asr"
-    )
-    def __init__(
-        self,
-        decoder: torch.nn.Module,
-        *,
-        nemo_asr: InjectedNemoModule = INJECTED,
-    ):
-        super().__init__()
-        self.decoder = decoder
-        self.active_fitering = isinstance(
-            decoder,
-            (
-                nemo_asr.modules.rnnt.RNNTDecoderJoint,
-                nemo_asr.modules.rnnt.RNNTDecoder,
-            ),
-        )
-
-    def _infer_batch_size(self, args, kwargs):
-        """Infer batch size from the first Tensor found.
-
-        This avoids relying on positional conventions.
-        """
-        for v in args:
-            if torch.is_tensor(v):
-                return v.shape[0], v
-        for v in kwargs.values():
-            if torch.is_tensor(v):
-                return v.shape[0], v
-        raise RuntimeError("Cannot infer batch size: no Tensor inputs found")
-
-    @property
-    def input_names(self):
-        if self.active_fitering:
-            return [
-                name
-                for name in self.decoder.input_names
-                if name != self.FILTER_ARGUMENT
-            ]
-        return self.decoder.input_names
-
-    @property
-    def output_names(self):
-        def rename_state(name: str) -> str:
-            if name == "states":
-                return "out_states"
-            return name
-
-        if self.active_fitering:
-            return [
-                rename_state(_)
-                for _ in self.decoder.output_names
-                if _ != self.FILTER_OUTPUT
-            ]
-        return self.decoder.output_names
-
-    @property
-    def index_arg_to_remove(self) -> int:
-        if self.active_fitering:
-            for idx, name in enumerate(self.decoder.input_names):
-                if name == self.FILTER_ARGUMENT:
-                    return idx
-        raise RuntimeError(
-            f"Cannot find argument named {self.FILTER_ARGUMENT} to remove"
-        )
-
-    @property
-    def index_output_to_remove(self) -> int:
-        if self.active_fitering:
-            for idx, name in enumerate(self.decoder.output_names):
-                if name == self.FILTER_OUTPUT:
-                    return idx
-        raise RuntimeError(
-            f"Cannot find output named {self.FILTER_OUTPUT} to remove"
-        )
-
-    def input_example(self, *args, **kwargs):
-        if not self.active_fitering:
-            return self.decoder.input_example(*args, **kwargs)
-        return self.filter_original_input_example(
-            self.decoder.input_example(*args, **kwargs)
-        )
-
-    def filter_original_input_example(
-        self, inputs: T.List[torch.Tensor]
-    ) -> T.List[torch.Tensor]:
-        """Filter out target_length from inputs."""
-        filtered_inputs = []
-        for name, tensor in zip(self.decoder.input_names, inputs):
-            if name != self.FILTER_ARGUMENT:
-                filtered_inputs.append(tensor)
-        return filtered_inputs
-
-    def forward(self, *args, **kwargs):
-        if not self.active_fitering:
-            return self.decoder(*args, **kwargs)
-
-        assert self.FILTER_ARGUMENT not in kwargs
-        batch_size, ref_tensor = self._infer_batch_size(args, kwargs)
-
-        target_length = torch.ones(
-            (batch_size, 1),
-            device=ref_tensor.device,
-            dtype=ref_tensor.dtype,
-        )
-        to_rm_in_idx = self.index_arg_to_remove
-        if len(args) > to_rm_in_idx:
-            args = list(args)
-            args.insert(to_rm_in_idx, target_length)
-            args = tuple(args)
-        else:
-            kwargs = dict(kwargs)
-            kwargs[self.FILTER_ARGUMENT] = target_length
-
-        outs = self.decoder(*args, **kwargs)
-
-        # If decoder returns multiple outputs:
-        # (logits, target_length, *states)
-        # Drop ONLY the target_length (2nd output)
-        to_rm_out_idx = self.index_output_to_remove
-        return tuple(
-            list(outs[:to_rm_out_idx]) + list(outs[to_rm_out_idx + 1 :])
-        )
-
-    def __getattr__(self, name):
-        try:
-            return super().__getattr__(name)
-        except AttributeError:
-            return getattr(self.decoder, name)
-
-
-class CollapseBatchDimWrapper(torch.nn.Module):
-    """Wrap a NeMo exportable subnet to remove batch from its interface.
-
-    - Accepts batch-less inputs and reconstructs batch=1 internally.
-    - Hides batch-only length-like inputs (e.g., 'length', 'target_length').
-    - Squeezes batch dim from tensor outputs and hides length-like outputs.
-
-    Notes:
-        - Requires the per-input dynamic axes mapping from NeMo
-          (as returned by `dynamic_shapes_for_export`).
-    """
-
-    LENGTH_INPUTS = {
-        "length",
-        "target_length",
-        "processed_length",
-        "audio_signal_length",
-    }
-    LENGTH_OUTPUTS = {
-        "encoded_lengths",
-        "prednet_lengths",
-        "length",
-        "processed_length",
-        "audio_signal_length",
-        "input_length",
-    }
-
-    def __init__(
-        self,
-        module: torch.nn.Module,
-        sym_dynamic_axes: T.Dict[str, T.Dict[int, str]],
-    ):
-        super().__init__()
-        self.module = module
-        # Mapping: input_name -> {axis_index: symbol like 'B','T','A',...}
-        self._sym_dynamic_axes = sym_dynamic_axes or {}
-        # Determine the external interface names (filtered)
-        self._orig_input_names = list(getattr(module, "input_names", []))
-        self._orig_output_names = list(getattr(module, "output_names", []))
-        def _is_len_name(n: str) -> bool:
-            nl = n.lower()
-            return (
-                nl in self.LENGTH_INPUTS
-                or nl in self.LENGTH_OUTPUTS
-                or ("length" in nl)
-            )
-
-        self._ext_input_names = [
-            n for n in self._orig_input_names if not _is_len_name(n)
-        ]
-        self._ext_output_names = [
-            n for n in self._orig_output_names if not _is_len_name(n)
-        ]
-
-        # Build a full-axes spec including static dims based on an input_example
-        self._full_axes_spec: T.Dict[str, T.List[str]] = {}
-        inner_ex = None
-        try:
-            ex = getattr(self.module, "input_example", None)
-            if ex is not None:
-                inner_ex = ex(max_batch=1)
-        except Exception:  # noqa: BLE001 - best effort only
-            inner_ex = None
-        # Try to use richer axis info from input_types when present
-        input_types = getattr(self.module, "input_types", None)
-
-        def _axis_kind_to_sym(kind: T.Any, default: str = "X") -> str:
-            name = None
-            if hasattr(kind, "name"):
-                name = str(kind.name).lower()
-            else:
-                name = str(kind).lower()
-            name_map = {
-                "batch": "B",
-                "time": "T",
-                "sequence": "S",
-                "stream": "S",
-                "length": "T",
-                "channel": "C",
-                "feature": "F",
-                "height": "H",
-                "width": "W",
-            }
-            return name_map.get(name, default)
-
-        for idx, name in enumerate(self._orig_input_names):
-            axes_dict = self._sym_dynamic_axes.get(name, {})
-            rank = 0
-            if (
-                inner_ex is not None
-                and idx < len(inner_ex)
-                and torch.is_tensor(inner_ex[idx])
-            ):
-                rank = inner_ex[idx].dim()
-            else:
-                rank = (max(axes_dict.keys()) + 1) if axes_dict else 1
-            # Start from input_types axis kinds if available
-            spec: T.List[str]
-            ntype = None
-            if isinstance(input_types, dict):
-                ntype = input_types.get(name)
-            if ntype is not None and hasattr(ntype, "axes"):
-                axes_kinds = list(getattr(ntype, "axes", ()))
-                spec = [_axis_kind_to_sym(getattr(k, "kind", k)) for k in axes_kinds]
-                # If input_types and observed rank disagree, pad/truncate
-                if len(spec) < rank:
-                    spec = spec + ["X"] * (rank - len(spec))
-                elif len(spec) > rank:
-                    spec = spec[:rank]
-            else:
-                spec = ["X"] * rank
-                for ax, sym in axes_dict.items():
-                    if 0 <= ax < rank:
-                        spec[ax] = sym
-            # Ensure batch axis presence when reasonable
-            if "B" not in spec and not _is_len_name(name) and len(spec) >= 1:
-                spec[0] = "B"
-            self._full_axes_spec[name] = spec
-
-        # Precompute collapsed axes map for external inputs using full spec
-        self._collapsed_axes = collapse_dynamic_axes_mapping(
-            self._sym_dynamic_axes, self._ext_input_names, full_axes_by_name=self._full_axes_spec
-        )
-
-    @property
-    def input_names(self) -> T.List[str]:
-        return self._ext_input_names
-
-    @property
-    def output_names(self) -> T.List[str]:
-        return self._ext_output_names
-
-    def dynamic_shapes_for_export(self, *args, **kwargs):
-        # Return collapsed mapping for external inputs
-        return self._collapsed_axes
-
-    def input_example(self, max_batch: int = 1):
-        # Build an example matching external interface (no batch, no lengths)
-        ex = getattr(self.module, "input_example", None)
-        inner = ex(max_batch=max_batch) if ex is not None else None
-        if inner is None:
-            # Synthesize minimal example using full axis spec
-            default_time = 16000
-            time_like = {"T", "A", "R", "S", "OUT"}
-            example: T.List[torch.Tensor] = []
-            for name in self._ext_input_names:
-                spec = self._full_axes_spec.get(name, ["X"])  # original ranks
-                # Collapsed shape skips 'B'
-                shape: T.List[int] = []
-                for sym in spec:
-                    if sym == "B":
-                        continue
-                    shape.append(default_time if sym in time_like else 1)
-                if not shape:
-                    shape = [default_time]
-                example.append(torch.zeros(*shape, dtype=torch.float32))
-            return example
-
-        # Map inner (orig) inputs to external list: drop length-like and squeeze B
-        out_examples: T.List[T.Any] = []
-        for name, tensor in zip(self._orig_input_names, inner):
-            if ("length" in name.lower()) or name in self.LENGTH_INPUTS:
-                continue
-            if torch.is_tensor(tensor):
-                axes = self._sym_dynamic_axes.get(name, {})
-                # squeeze all B axes (commonly axis 0)
-                # Apply in decreasing order to keep indices valid
-                b_axes = sorted(
-                    [
-                        i
-                        for i, s in enumerate(
-                            self._full_axes_spec.get(name, [])
-                        )
-                        if s == "B"
-                    ],
-                    reverse=True,
-                )
-                t = tensor
-                for ax in b_axes:
-                    if 0 <= ax < t.dim() and t.size(ax) == 1:
-                        t = t.squeeze(dim=ax)
-                # Reduce any extra dims to the expected collapsed rank by taking
-                # the first item along leading extra axes (keep last dims)
-                expected_rank = sum(
-                    1 for s in self._full_axes_spec.get(name, []) if s != "B"
-                )
-                while torch.is_tensor(t) and t.dim() > max(1, expected_rank):
-                    t = t.select(dim=0, index=0)
-                out_examples.append(t)
-            else:
-                out_examples.append(tensor)
-        return out_examples
-
-    def _infer_time_from_visible_inputs(self, args: T.Sequence[T.Any]) -> int:
-        # Infer time length from full axis spec (works even if time is static)
-        name_to_arg: T.Dict[str, T.Any] = {
-            n: a for n, a in zip(self._ext_input_names, args)
-        }
-        time_like = {"T", "A", "R", "S", "OUT"}
-        times: T.List[int] = []
-        for name, tensor in name_to_arg.items():
-            if not torch.is_tensor(tensor):
-                continue
-            spec = self._full_axes_spec.get(name, [])
-            collapsed_index = 0
-            for sym in spec:
-                if sym == "B":
-                    continue
-                if sym in time_like and collapsed_index < tensor.dim():
-                    times.append(int(tensor.shape[collapsed_index]))
-                collapsed_index += 1
-        return max(times) if times else 1
-
-    def _synthesize_length_tensor(
-        self,
-        ref_device: torch.device,
-        ref_dtype: torch.dtype,
-        length: int,
-        name: str,
-    ) -> torch.Tensor:
-        if name == "target_length":
-            # keep an extra last dim to match common expectation (B,1)
-            return torch.tensor([[length]], device=ref_device, dtype=torch.long)
-        return torch.tensor([length], device=ref_device, dtype=torch.long)
-
-    def forward(self, *args, **kwargs):
-        # args correspond to external inputs in order self._ext_input_names
-        # Rebuild inner args matching original interface
-        assert not kwargs, (
-            "CollapseBatchDimWrapper expects positional args only"
-        )
-        visible = list(args)
-        full: T.List[T.Any] = []
-
-        # Determine a reference device/dtype and time
-        ref_tensor = next((a for a in visible if torch.is_tensor(a)), None)
-        ref_device = (
-            ref_tensor.device
-            if torch.is_tensor(ref_tensor)
-            else torch.device("cpu")
-        )
-        ref_dtype = (
-            ref_tensor.dtype if torch.is_tensor(ref_tensor) else torch.float32
-        )
-        time_len = self._infer_time_from_visible_inputs(visible)
-
-        # map from visible index to name for convenience
-        vis_iter = iter(visible)
-        for name in self._orig_input_names:
-            if ("length" in name.lower()) or name in self.LENGTH_INPUTS:
-                full.append(
-                    self._synthesize_length_tensor(
-                        ref_device, ref_dtype, time_len, name
-                    )
-                )
-                continue
-            val = next(vis_iter)
-            if torch.is_tensor(val):
-                axes = self._sym_dynamic_axes.get(name, {})
-                # Reduce any unexpected extra dims to match collapsed spec
-                expected_rank = sum(
-                    1 for s in self._full_axes_spec.get(name, []) if s != "B"
-                )
-                t = val
-                while t.dim() > max(1, expected_rank):
-                    t = t.select(dim=0, index=0)
-                # Insert batch=1 at each 'B' axis position
-                b_axes = sorted(
-                    [
-                        i
-                        for i, s in enumerate(
-                            self._full_axes_spec.get(name, [])
-                        )
-                        if s == "B"
-                    ]
-                )  # asc order
-                # Adjust for growing rank when inserting
-                offset = 0
-                for ax in b_axes:
-                    t = t.unsqueeze(dim=ax + offset)
-                    offset += 1
-                full.append(t)
-            else:
-                full.append(val)
-
-        outs = self.module(*tuple(full))
-        if not isinstance(outs, tuple):
-            outs = (outs,)
-
-        # Drop length-like outputs and squeeze batch for tensors
-        keep_indices = [
-            i
-            for i, n in enumerate(self._orig_output_names)
-            if n not in self.LENGTH_OUTPUTS
-        ]
-        proc: T.List[T.Any] = []
-        for i in keep_indices:
-            o = outs[i]
-            if torch.is_tensor(o):
-                # heuristic: squeeze leading batch dim when size==1
-                if o.dim() > 0 and o.size(0) == 1:
-                    o = o.squeeze(0)
-            proc.append(o)
-        return tuple(proc)
-
-    def __getattr__(self, name):
-        try:
-            return super().__getattr__(name)
-        except AttributeError:
-            return getattr(self.module, name)
-
-
-def use_pytorch_sdpa(model: torch.nn.Module):
-    """Modify the model to use PyTorch sdpa implementations where applicable.
-
-    This leverage attention modules set in NeMo with
-    specific use_pytorch_sdpa flag.
-    """
-    # pylint: disable=import-outside-toplevel
-    from nemo.collections.asr.parts.submodules.multi_head_attention import (
-        MultiHeadAttention,
-    )
-
-    for module in model.modules():
-        if isinstance(module, MultiHeadAttention):
-            module.use_pytorch_sdpa = True
-
-
-@require_extra_decorator(extra=T2NExtra.NEMO_TRACT, module="questionary")
-def ask_model_selector(
-    pretrained_model_info_list,
-    *,
-    questionary: InjectedQuestionaryModule = INJECTED,
-    **kwargs,
-):
-    # create the question object
-    question = questionary.select(
-        " ∵ What model do you want to export ? "
-        "(those starting with nvidia/ are from 🤗Hub)",
-        qmark="",
-        choices=[
-            questionary.Choice(
-                title=pretrained_model_info.pretrained_model_name,
-                description=pretrained_model_info.description,
-            )
-            for pretrained_model_info in pretrained_model_info_list
-        ],
-        use_jk_keys=False,
-        use_search_filter=True,
-        **kwargs,
-    )
-
-    # prompt the user for an answer
-    return question.ask()
-
-
-def nemo_asr_hg_list(huggingface_hub: HuggingFaceHubModule):
-    """Return the list of available NeMo ASR models from HuggingFace."""
-    hugging_face_hub_model_list = []
-
-    # Query HF for NVIDIA ASR models that declare NeMo compatibility
-    models = huggingface_hub.list_models(
-        author="nvidia",
-        task="automatic-speech-recognition",
-        library="nemo",
-        full=True,
-    )
-
-    for m in models:
-        # Skip archived / gated / unusable entries early
-        if getattr(m, "gated", False):
-            continue
-
-        tags = set(m.tags or [])
-
-        # Heuristic: ensure this is an ASR *model*,
-        # not speech translation / TTS / etc.
-        if "automatic-speech-recognition" not in tags:
-            continue
-
-        jtags = "tags:" + ", ".join(sorted(tags))
-
-        def last_modified_str(dt: datetime.datetime) -> str:
-            fmt = "%Y-%m-%d"
-            return f"last modified: {dt.strftime(fmt)}"
-
-        hugging_face_hub_model_list.append(
-            HGModelInfo(
-                pretrained_model_name=m.modelId,
-                organization="nvidia",
-                description="HuggingFace (compatibility only guessed): "
-                + (m.cardData.get("description") if m.cardData else jtags)
-                + last_modified_str(m.lastModified),
-                tags=sorted(tags),
-                pipeline_tag=m.pipeline_tag,
-                library_name=m.library_name,
-                sha=m.sha,
-                last_modified=m.lastModified,
-            )
-        )
-
-    # Sort to match NeMo-style deterministic output
-    hugging_face_hub_model_list.sort(
-        key=lambda x: x.last_modified, reverse=True
-    )
-
-    return hugging_face_hub_model_list
-
-
-@dataclass(frozen=True)
-class HGModelInfo:
-    pretrained_model_name: str
-    organization: str
-    description: str
-    tags: T.List[str]
-    pipeline_tag: str
-    library_name: str
-    sha: str
-    last_modified: datetime.datetime
-
-
-@require_extra_decorator(
-    extra=T2NExtra.NEMO_TRACT, module="nemo.collections.asr", kw="nemo_asr"
-)
-@require_extra_decorator(extra=T2NExtra.NEMO_TRACT, module="huggingface_hub")
-def load_asr_model_from_nemo_slug(
-    model_slug: str,
-    *,
-    nemo_asr: InjectedNemoModule = INJECTED,
-    huggingface_hub: InjectedHuggingFaceHubModule = INJECTED,
-):
-    """Load a NeMo ASR model from a given model slug."""
-    # pylint: disable=import-outside-toplevel
-    from huggingface_hub import errors
-
-    try:
-        asr_model = nemo_asr.models.ASRModel.from_pretrained(
-            model_name=model_slug, map_location=torch.device("cpu")
-        )
-    except (errors.RepositoryNotFoundError, FileNotFoundError):
-        LOGGER.error("Could not find model with slug: %s", model_slug)
-        if model_slug != "*":
-            while True:
-                resp = (
-                    input("Do you want to list available models? (y/n): ")
-                    .strip()
-                    .lower()
-                )
-                if resp in ("y", "n"):
-                    break
-            if resp == "n":
-                LOGGER.info("User chose not to list available models. Exiting.")
-                sys.exit(1)
-
-        available_models = (
-            nemo_asr_hg_list(huggingface_hub)
-            + nemo_asr.models.ASRModel.list_available_models()
-        )
-        model_slug = ask_model_selector(available_models)
-        LOGGER.info("selected model slug: %s", model_slug)
-        asr_model = nemo_asr.models.ASRModel.from_pretrained(
-            model_name=model_slug,
-            map_location=torch.device("cpu"),
-        )
-    return asr_model
-
-
-def main():
-    init_log()
-    args = parser_cli()
-    log_level = logging.INFO
-    if args.verbose:
-        log_level = logging.DEBUG
-    set_lib_log_level(log_level)
-    export_dir = Path(args.export_dir)
-    assert not export_dir.exists(), f"export_dir '{export_dir}' must not exist"
-    export_dir.mkdir(parents=True, exist_ok=False)
-
-    handler = logging.FileHandler(export_dir / "nemo_tract_export.log")
-    handler.setFormatter(
-        logging.Formatter(
-            "%(asctime)s,%(msecs)d %(levelname)-8s "
-            "[%(filename)s:%(lineno)d] %(message)s",
-            "%Y-%m-%d:%H:%M:%S",
-        )
-    )
-    logging.getLogger().addHandler(handler)
-    LOGGER.info("started nemo_tract export with args: %s", args)
-    # ensure that the model is loaded on CPU
-    asr_model = load_asr_model_from_nemo_slug(args.model_slug)
-
-    if args.force_sdpa_pytorch:
-        use_pytorch_sdpa(asr_model)
-    asr_model.eval()
-
-    if args.data_type == "float16":
-        asr_model = asr_model.half()
-        asr_model.preprocessor.to(dtype=torch.float32)
-
-    if args.data_type in ["float16", "mixed"] and hasattr(
-        asr_model, "preprocessor"
-    ):
-        asr_model.preprocessor = WrapPreprocessorCast(
-            asr_model.preprocessor, dtype=torch.float16
-        )
-
-    if (
-        isinstance(args.tract_check_io_tolerance, str)
-        and args.tract_check_io_tolerance != "skip"
-    ):
-        args.tract_check_io_tolerance = TractCheckTolerance(
-            args.tract_check_io_tolerance
-        )
-
-    inference_target = setup_inference_target_from_cli_args(args)
-
-    with (export_dir / "export_config.json").open("w", encoding="utf8") as fh:
-        json.dump(
-            {
-                k: str(v) if isinstance(v, Path) else v
-                for k, v in vars(args).items()
-            },
-            fh,
-            indent=2,
-        )
-
-    def call_export(float_dtype=torch.float32):
-        export_nemo_asr_model(
-            asr_model,
-            inference_target,
-            export_dir,
-            nnef_variable_naming_scheme=VariableNamingScheme(
-                args.naming_scheme
-            ),
-            compress_registry=args.compress_registry,
-            compress_method=args.compress_method,
-            skip_preprocessor=args.skip_preprocessor,
-            split_joint_decoder=args.split_joint_decoder,
-            extra_cfg={"pretrained_name": args.model_slug},
-            float_dtype=float_dtype,
-            dump_checked_io=args.dump_checked_io,
-            collapse_batch_dim=args.collapse_batch_dim,
-        )
-
-    if args.data_type == "mixed":
-        try:
-            # pylint: disable=import-outside-toplevel
-            from torch import autocast
-
-            LOGGER.info("exporting with mixed precision using autocast")
-            LOGGER.warning(
-                "mixed precision export is experimental "
-                "(not supported by tract)"
-            )
-            with autocast(device_type="cpu", dtype=torch.float16):
-                call_export(float_dtype=torch.float16)
-        except ImportError as ie:
-            raise ImportError(
-                "To use mixed precision export please install recent torch"
-            ) from ie
-    else:
-        call_export(
-            float_dtype=torch.float16
-            if args.data_type == "float16"
-            else torch.float32
-        )
-
-
-if __name__ == "__main__":
-    main()
